move install_on_import to pyodide, add package.json schema for pyodideKernel
