+++ jupyterlite_pyodide_kernel/tests/test_metadata.py
@@ -1,3 +1,4 @@
+"""verify the labextension metadata is correct"""
 import jupyterlite_pyodide_kernel
