+++ jupyterlite_pyodide_kernel/tests/test_repo.py
@@ -1,0 +1,43 @@
+"""Validate the integrity of the repo.
+
+Assumes:
+- a ``git`` checkout
+- the labextension has already been built with ``jlpm build``
+"""
+import pytest
+import json
+from pathlib import Path
+
+from jupyterlite_core.constants import UTF8
+from jupyterlite_pyodide_kernel.constants import PYODIDE_VERSION
+from .conftest import IN_TREE_EXTENSION, HERE
+
+PACKAGES = HERE / "../../packages"
+
+if not IN_TREE_EXTENSION.exists():  # pragma: no cover
+    pytest.skip("not a source build, skipping repo tests", allow_module_level=True)
+
+
+def test_pyodide_version():
+    kernel_pkg = PACKAGES / "pyodide-kernel"
+    kernel_pkg_json = kernel_pkg / "package.json"
+    kernel_pkg_data = json.loads(kernel_pkg_json.read_text(**UTF8))
+    assert (
+        kernel_pkg_data["devDependencies"]["pyodide"] == PYODIDE_VERSION
+    ), f"{kernel_pkg_data} pyodide devDependency is not {PYODIDE_VERSION}"
+
+
+@pytest.fixture
+def the_default_pyodide_url():
+    return f"https://cdn.jsdelivr.net/pyodide/v{PYODIDE_VERSION}/full/pyodide.js"
+
+
+@pytest.mark.parametrize(
+    "pkg_path",
+    [
+        "pyodide-kernel-extension/schema/kernel.v0.schema.json",
+        "pyodide-kernel-extension/src/index.ts",
+    ],
+)
+def test_pyodide_url(pkg_path: Path, the_default_pyodide_url: str):
+    assert the_default_pyodide_url in (PACKAGES / pkg_path).read_text(**UTF8)
