+++ package.json
@@ -42,7 +42,7 @@
     "prettier": "jlpm prettier:base --write --list-different",
     "quickstart": "npm run setup:py && jlpm && jlpm deduplicate && jlpm clean:all && jlpm bootstrap && jlpm lint && jlpm build:prod && jlpm dist && jlpm docs && jlpm test",
     "serve": "cd build && python -m http.server -b 127.0.0.1",
-    "setup:py": "python -m pip install -e \".[dev,lint,test,docs]\"",
+    "setup:py": "python -m pip install -r requirements-dev.txt",
     "test:py": "pytest",
     "test": "jlpm test:py",
     "watch": "lerna run watch"
