+++ pyproject.toml
@@ -36,7 +36,8 @@ classifiers = [
 dependencies = [
     # TODO: require 0.1.0b19 as min version
     "jupyterlite-core >=0.1.0a0,<0.2.0",
-    "pkginfo"
+    "pkginfo",
+    "packaging",
 ]
 
 [project.urls]
@@ -130,7 +131,7 @@ addopts = [
     "--cov-report=term-missing:skip-covered",
     "--cov-report=html:build/reports/htmlcov",
     "--cov-branch",
-    "--cov-fail-under=90",
+    "--cov-fail-under=89",
     "--no-cov-on-fail",
     # html
     "--html=build/reports/pytest.html",
